@@ -9,8 +9,8 @@
 package com.example.samplestickerapp;
 
 import android.content.Context;
-import android.support.annotation.Nullable;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.Nullable;
+import androidx.recyclerview.widget.RecyclerView;
 import android.util.AttributeSet;
 
 public class BottomFadingRecyclerView extends RecyclerView {
