@@ -8,7 +8,7 @@
 
 package com.example.samplestickerapp;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.text.TextUtils;
 import android.util.JsonReader;
 
