@@ -11,7 +11,7 @@ package com.example.samplestickerapp;
 import android.content.Context;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.text.TextUtils;
 
 import com.facebook.animated.webp.WebPImage;
