@@ -8,7 +8,7 @@
 
 package com.example.samplestickerapp;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 import com.facebook.drawee.view.SimpleDraweeView;
